+++ bindings/nodejs/types/client/outputBuilderOptions/aliasOutputParams.ts
@@ -1,13 +1,14 @@
-import type { FeatureTypes, HexEncodedString } from '@iota/types';
-import type { IBasicOutputBuilderParams } from './basicOutputParams';
+import { HexEncodedString } from '@iota/types';
+import { Feature } from '../../../lib';
+import type { BasicOutputBuilderParams } from './basicOutputParams';
 
 /**
  * Options for building an Alias Output
  */
-export interface IAliasOutputBuilderParams extends IBasicOutputBuilderParams {
+export interface AliasOutputBuilderParams extends BasicOutputBuilderParams {
     aliasId: HexEncodedString;
     stateIndex?: number;
     stateMetadata?: HexEncodedString;
     foundryCounter?: number;
-    immutableFeatures?: FeatureTypes[];
+    immutableFeatures?: Feature[];
 }