+++ bindings/python/examples/client/02_generate_addresses.py
@@ -1,12 +1,21 @@
 from iota_sdk import Client, MnemonicSecretManager, CoinType
 
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+node_url = os.environ.get('NODE_URL', 'https://api.testnet.shimmer.network')
+
 # Create a Client instance
-client = Client(nodes=['https://api.testnet.shimmer.network'])
+client = Client(nodes=[node_url])
+
+if 'NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1' not in os.environ:
+    print(".env mnemonic is undefined, see .env.example")
+    sys.exit(1)
 
 # In this example we will create addresses from a mnemonic
-
-secret_manager = MnemonicSecretManager(
-    "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river")
+secret_manager = MnemonicSecretManager(os.environ['NON_SECURE_USE_OF_DEVELOPMENT_MNEMONIC_1'])
 
 # Generate public address with default account index and range.
 addresses = client.generate_addresses(secret_manager)